+++ CHANGELOG.md
@@ -6,6 +6,7 @@ v1.8.1 (??.06.2022)
 * removed all anient model classes
 * Fix filter classes
 * Fix fatal error in stadium list
+* MatchNoteMarker renders match data again
 
 v1.8.0 (19.06.2022)
 * all classes beside from models are moved to PSR-4 namespace